--- WcaOnRails/app/webpacker/components/contexts.jsx
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const DelegateMattersContext = createContext(false);
