+++ WcaOnRails/app/webpacker/lib/contexts.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const DelegateMattersContext = createContext(false);
+
+export default DelegateMattersContext;
